Fix macro type collission bug
Since we have properties and values with the same name macro
incorrectly generated types for the structs and enums. In this commit
we denote Leafs as Keywords or Values and Named as Other properties or
generated structs or enums. This prevents name collision, since using
this information we can add prefix `values::` to correct type to
denote that this is value and not a property.